backend/requirements.txt
# AccuDesign — Backend Dependencies (Phase 4)
# Install: pip install -r requirements.txt

# ── Web Framework ─────────────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9         # Required for UploadFile / FormData

# ── B-Rep Geometry Engine ─────────────────────────────────────────────────────
cadquery==2.4.0                  # CadQuery wraps OpenCASCADE (OCCT) for B-Rep analysis

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
pydantic==1.10.21               # v1 — @validator syntax; v2 requires @field_validator
httpx==0.27.0                    # Async HTTP for price APIs + exchange rate APIs
fpdf2==2.7.9                     # PDF quote generation (ACCU DESIGN format)
numpy==1.26.4                    # Array ops for geometry utils

# ── AI / PDF Analysis ────────────────────────────────────────────────────────
google-generativeai>=0.8.0       # Google Gemini for PDF engineering drawing analysis
